requirements.txt
Django==2.1.1
django-bootstrap4==0.0.7
djangorestframework==3.8.2
pytz==2018.5
